Added user configuration files snapshot endpoints
